+++ server/blockchain/src/database/projections/mod.rs
@@ -1,4 +1,5 @@
 use chrono::{DateTime, Utc};
+use num_bigint::BigInt;
 
 pub mod asset;
 pub mod deposit;
@@ -7,7 +8,7 @@ pub mod valut;
 pub mod withdraw;
 
 pub trait Confirmable {
-    fn set(&mut self, confirmations: usize);
+    fn set(&mut self, confirmations: BigInt);
     fn is_confirmed(&self) -> bool;
 }
